# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "Server API to manage system operations and resources."
  version: "1.0.0"
  title: "API seu estoque"
  contact:
    email: "fagner.m.carvalho@gmail.com"
host: "virtserver.swaggerhub.com"
basePath: "/Fagner-Carvalho/seu-estoque-API/1.0.0"
...
schemes:
- "https"
- "http"
paths:
  /user:
    get:
      tags:
      - "user"
      summary: "List all users"
      description: ""
      operationId: "ListUsers"
      parameters:
      - name: "page"
        in: "query"
        description: "The page number"
        required: false
        type: "integer"
        minimum: 1
      - name: "limit"
        in: "query"
        description: "The numbers of items to return"
        required: false
        type: "integer"
        minimum: 1
      - name: "sort"
        in: "query"
        description: "The field to be ordered"
        required: false
        type: "string"
      - name: "order"
        in: "query"
        description: "The sorting type (ascending or descending)"
        required: false
        type: "string"
        enum:
        - "asc"
        - "desc"
      - name: "q"
        in: "query"
        description: "The text to filter user"
        required: false
        type: "string"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/User"
        "400":
          description: "failed operation"
      x-swagger-router-controller: "User"
    post:
      tags:
      - "user"
      summary: "Add a new user"
      description: ""
      operationId: "addUser"
      parameters:
      - in: "body"
        name: "body"
        description: "User object that needs to be added"
        required: true
        schema:
          $ref: "#/definitions/User"
      responses:
        "200":
          description: "successful operation"
        "400":
          description: "failed operation"
      x-swagger-router-controller: "User"
  /user/{userId}:
    get:
      tags:
      - "user"
      summary: "Find user by ID"
      description: "Returns a single user"
      operationId: "getUserById"
      parameters:
      - name: "userId"
        in: "path"
        description: "ID of user to return"
        required: true
        type: "string"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/User"
        "400":
          description: "failed operation"
        "404":
          description: "User not found"
      x-swagger-router-controller: "User"
    put:
      tags:
      - "user"
      summary: "Updates a user with form data"
      description: "Returns a single user"
      operationId: "updateUserWithForm"
      parameters:
      - name: "userId"
        in: "path"
        description: "ID of user that needs to be updated"
        required: true
        type: "string"
      responses:
        "200":
          description: "successful operation"
        "400":
          description: "failed operation"
        "404":
          description: "User not found"
      x-swagger-router-controller: "User"
    delete:
      tags:
      - "user"
      summary: "Deletes a user"
      description: ""
      operationId: "deleteUser"
      parameters:
      - name: "userId"
        in: "path"
        description: "ID of user that needs to be updated"
        required: true
        type: "string"
      responses:
        "200":
          description: "successful operation"
        "400":
          description: "failed operation"
        "404":
          description: "User not found"
      x-swagger-router-controller: "User"
...
  /unitMeasure:
    get:
      tags:
      - "unitMeasure"
      summary: "List all unitMeasure"
      description: ""
      operationId: "ListUnitMeasure"
      parameters:
      - name: "page"
        in: "query"
        description: "The page number"
        required: false
        type: "integer"
        minimum: 1
      - name: "limit"
        in: "query"
        description: "The numbers of items to return"
        required: false
        type: "integer"
        minimum: 1
      - name: "sort"
        in: "query"
        description: "The field to be ordered"
        required: false
        type: "string"
      - name: "order"
        in: "query"
        description: "The sorting type (ascending or descending)"
        required: false
        type: "string"
        enum:
        - "asc"
        - "desc"
      - name: "q"
        in: "query"
        description: "The text to filter unitMeasure"
        required: false
        type: "string"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/UnitMeasure"
        "400":
          description: "failed operation"
      x-swagger-router-controller: "UnitMeasure"
    post:
      tags:
      - "unitMeasure"
      summary: "Add a new unitMeasure"
      description: ""
      operationId: "addUnitMeasure"
      parameters:
      - in: "body"
        name: "body"
        description: "UnitMeasure object that needs to be added"
        required: true
        schema:
          $ref: "#/definitions/UnitMeasure"
      responses:
        "200":
          description: "successful operation"
        "400":
          description: "failed operation"
      x-swagger-router-controller: "UnitMeasure"
  /unitMeasure/{unitMeasureId}:
    get:
      tags:
      - "unitMeasure"
      summary: "Find unitMeasure by ID"
      description: "Returns a single unitMeasure"
      operationId: "getUnitMeasureById"
      parameters:
      - name: "unitMeasureId"
        in: "path"
        description: "ID of unitMeasure to return"
        required: true
        type: "string"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/UnitMeasure"
        "400":
          description: "failed operation"
        "404":
          description: "UnitMeasure not found"
      x-swagger-router-controller: "UnitMeasure"
    put:
      tags:
      - "unitMeasure"
      summary: "Updates a unitMeasure with form data"
      description: "Returns a single unitMeasure"
      operationId: "updateUnitMeasureWithForm"
      parameters:
      - name: "unitMeasureId"
        in: "path"
        description: "ID of unitMeasure that needs to be updated"
        required: true
        type: "string"
      responses:
        "200":
          description: "successful operation"
        "400":
          description: "failed operation"
        "404":
          description: "UnitMeasure not found"
      x-swagger-router-controller: "UnitMeasure"
    delete:
      tags:
      - "unitMeasure"
      summary: "Deletes a unitMeasure"
      description: ""
      operationId: "deleteUnitMeasure"
      parameters:
      - name: "unitMeasureId"
        in: "path"
        description: "ID of unitMeasure that needs to be updated"
        required: true
        type: "string"
      responses:
        "200":
          description: "successful operation"
        "400":
          description: "failed operation"
        "404":
          description: "UnitMeasure not found"
      x-swagger-router-controller: "UnitMeasure"
...
definitions:
  User:
    type: "object"
    required:
    - "email"
    - "name"
    properties:
      id:
        type: "string"
      name:
        type: "string"
      email:
        type: "string"
      password:
        type: "string"
      phone:
        type: "string"
      status:
        type: "boolean"
        description: "User Status"
        default: true
      createdAt:
        type: "string"
        format: "datetime"
      updatedAt:
        type: "string"
        format: "datetime"
      deletedAt:
        type: "string"
        format: "datetime"
    example:
      createdAt: "createdAt"
      password: "password"
      deletedAt: "deletedAt"
      phone: "phone"
      name: "name"
      id: "id"
      email: "email"
      status: true
      updatedAt: "updatedAt"
...
  UnitMeasure:
    type: "object"
    required:
    - "name"
    properties:
      id:
        type: "string"
      name:
        type: "string"
      abbreviation:
        type: "string"
      createdAt:
        type: "string"
        format: "datetime"
      updatedAt:
        type: "string"
        format: "datetime"
      deletedAt:
        type: "string"
        format: "datetime"
    example:
      createdAt: "createdAt"
      deletedAt: "deletedAt"
      name: "name"
      id: "id"
      abbreviation: "abbreviation"
      updatedAt: "updatedAt"
...